pyproject.toml
[project]
name = "payswitch-cs-dataml"
version = "0.1.0"
description = "PaySwitch Credit Scoring ML Training and Inference System"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.2,<3",
    "numpy>=1.26,<2",
    "pymupdf>=1.27.2.2",
    "azure-functions>=1.24.0",
    "azure-servicebus>=7.14.3",
    "azure-storage-blob>=12.28.0",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "xgboost>=3.2.0",
    "shap>=0.49.1",
    "optuna>=4.8.0",
    "mlflow>=3.10.1",
    "scikit-learn>=1.8.0",
    "lightgbm>=4.6.0",
    "azure-ai-ml>=1.32.0",
    "azure-identity>=1.25.3",
    "setuptools>=82.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
